refactor(radiobutton): move radio button as main component
