--- lib/src/RadioGroup/index.ts
@@ -1,2 +1,0 @@
-export * from "./RadioGroup";
-export * from "./RadioButton/RadioButton";
